Fixes issue 86. Gene shadow drawing problem was resolved. And also, Wig data uses int types for representing (start, end, width)

+++ utgb-core/src/main/java/org/utgenome/gwt/utgb/client/BrowserService.java
@@ -85,12 +85,13 @@ public interface BrowserService extends RemoteService {
 
 	public List<DatabaseEntry> getDBEntry(String dbGroup);
 
-	public List<WigGraphData> getWigDataList(String fileName, long windowWidth, ChrLoc location);
+	public List<WigGraphData> getWigDataList(String fileName, int windowWidth, ChrLoc location);
 
 	public List<SAMRead> getSAMReadList(String readFileName, String refSeqFileName);
 
 	public List<Gene> getBEDEntryList(String bedPath, ChrLoc location);
 
 	public List<SAMRead> querySAMReadList(String bamFileName, String indexFileName, String refSeqFileName, String rname, int start, int end);
+
 	public String getRefSeq(String refSeqFileName, String rname, int start, int end);
 }
