+++ utgb-core/src/main/java/org/utgenome/gwt/utgb/client/bio/CDS.java
@@ -40,13 +40,13 @@ public class CDS implements Serializable {
 	/**
 	 * 
 	 */
-	long start = -1;
-	long end = -1;
+	int start = -1;
+	int end = -1;
 
 	public CDS() {
 	}
 
-	public CDS(long start, long end) {
+	public CDS(int start, int end) {
 		this.start = start;
 		this.end = end;
 	}
@@ -58,19 +58,19 @@ public class CDS implements Serializable {
 			end += 1;
 	}
 
-	public long getStart() {
+	public int getStart() {
 		return start;
 	}
 
-	public void setStart(long start) {
+	public void setStart(int start) {
 		this.start = start;
 	}
 
-	public long getEnd() {
+	public int getEnd() {
 		return end;
 	}
 
-	public void setEnd(long end) {
+	public void setEnd(int end) {
 		this.end = end;
 	}
